+++ src/screens/tutorials/TimePickerScreen.js
@@ -1,43 +1,15 @@
-import React, { useEffect, useState } from 'react';
-import {View, Text, Image} from 'react-native';
+import React from 'react';
+import {View} from 'react-native';
 import {useSelector} from 'react-redux';
 
-import { TimePicker } from '../../components/tutorials/TimePicker';
 import { StepContentsComponent } from '../../components/tutorials/StepContentsComponent';
-import { PretendardedText } from '../../components/CustomComponent/PretendardedText';
 import { RoutineAddListComponent } from '../../components/RoutineAddList/RoutineAddListComponent';
-
-import { asPickerFormat } from '../../utils/tutorials/TutorialUtils';
-import { BUTTON_HEIGHT, VIEW_WIDTH } from '../../utils/tutorials/Values';
-
+import { RoutineTimePicker } from '../../components/RoutineTimePicker';
 import { Step } from '../../utils/tutorials/Step';
-
-import GreenStarImage from '../../assets/images/GreenStar.png';
 
 export const TimePickerScreen = () => {
   const textColor = useSelector((state) => state.tutorial.textColor);
-  const stepBottomContentOpacity = useSelector((state) => state.tutorial.stepBottomContentOpacity);
   const step = useSelector((state) => state.tutorial.step);
-  const [time, setTime] = useState(asPickerFormat());
-  const [routineTime, setroutineTime] = useState(0);
-  const [isValid, setisValid] = useState(true);
-  const startTime = useSelector((state) => state.user.wakeUpTime);
-  const finishTime = useSelector((state) => state.user.completeTime);
-
-  useEffect(() => {
-      var tempStartTime = new Date(startTime);
-      var startMinutes = tempStartTime.getHours() * 60 + tempStartTime.getMinutes();
-      var tempFinishTime = new Date(finishTime);
-      var finishMinutes = tempFinishTime.getHours() * 60 + tempFinishTime.getMinutes();
-
-      if(finishMinutes >= startMinutes){
-        setisValid(true);
-      }else{
-        setisValid(false);
-      }
-
-      setroutineTime(finishMinutes - startMinutes);
-  }, [startTime, finishTime]);
 
   return (
     <View style={{flex: 1, margin:15, justifyContent:'space-between', flexDirection:'column'}}>
@@ -49,23 +21,7 @@ export const TimePickerScreen = () => {
           step === Step.STEP_THREE ?
             <RoutineAddListComponent isTutorial={true}/>
           :
-          <View style={{position:'relative', width:'100%', height:'100%'}}>
-            <TimePicker
-              value={time}
-              onChange={setTime}
-              buttonHeight={BUTTON_HEIGHT}
-              visibleCount={5}
-            />
-            <View style={{height: 100, alignItems:'flex-end', justifyContent:'center', opacity:stepBottomContentOpacity, position:'absolute', right:0, bottom:0}}>
-              <PretendardedText style={{color:textColor, fontSize: 15, fontWeight: 500}}>나에게 주어진 총 루틴 시간은</PretendardedText>
-              <PretendardedText style={{color:textColor, fontSize: 15, fontWeight: 500, textAlign:'right'}}>기상시간 기준</PretendardedText>
-              <View style={{flexDirection:'row', justifyContent:'flex-end'}}>
-                <PretendardedText style={{color:isValid ? "#3CE3AC" : "#FF6056", fontSize: 15, fontWeight: 900}}>{isValid ? "+" : ""}{routineTime}</PretendardedText>
-                <PretendardedText style={{color:isValid ? "#3CE3AC" : "#FF6056", fontSize: 15, fontWeight: 900}}>분</PretendardedText>
-                <PretendardedText style={{color:textColor, fontSize: 15, fontWeight: 500}}> 입니다.</PretendardedText>
-              </View>
-            </View>
-          </View>
+          <RoutineTimePicker/>
         }
       </View>
     </View>
